+++ BidmadPluginSample/Assets/Bidmad/Scripts/ButtonEvent.cs
@@ -20,6 +20,11 @@ public class ButtonEvent : MonoBehaviour
         SceneManager.LoadScene("Reward");
     }
 
+    public void MoveRewardInterstitialScene()
+    {
+        SceneManager.LoadScene("RewardInterstitial");
+    }
+
     public void MoveMainScene()
     {
         SceneManager.LoadScene("Main");